Update README.md
Sahana Read this to understand how u can edit the code. Since ur VS Code is not working u can just edit the code. I will see the output. We have index.html and quiz.html file that we have to edit . index.html file is the first cover page of the website and the quiz.html is the actuall quiz. We have to share the work and makes changes for these to improve. After this we have to edit the quiz.js where we have to add all the movies and the posters will see that later